[FIX] Notice 도메인 매니저 pk 값 전달 추가
