+++ src/main/java/com/kidmily/algoga_server/notice/application/command/CreateNoticeCommand.java
@@ -5,6 +5,6 @@ import com.kidmily.algoga_server.notice.presentation.NoticeTagType;
 public record CreateNoticeCommand(
         String title,
         String content,
-        NoticeTagType noticeTagType
-) {
-}
+        NoticeTagType noticeTagType,
+        Long managerId // 🌟 매니저 ID 필드 추가
+) {}
